Modification Nester push main ?

+++ nester.py
@@ -4,6 +4,7 @@ from flask_sqlalchemy import SQLAlchemy
 import time
 import threading
 import datetime
+import socket
 
 app = Flask(__name__) # Instance de la classe Flask 
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///bdd.db' #URI de la bdd qui va être crée  
@@ -33,11 +34,11 @@ class HarvesterDetails(db.Model):
     agent_version = db.Column(db.String(50))
 
 # Création de la forme de la db 
-def create_tables():
-    db.create_all()
+#def create_tables():
+#   db.create_all()
 
-def setup():
-    create_tables()
+#def setup():
+#   create_tables()
                     
 def manage_status_v2():
     with app.app_context():
@@ -54,9 +55,10 @@ def manage_status_v2():
                         print(float(client.last_request))
                         print(now - float(client.last_request))
                         client.statut = "Disconnected"
-                    elif now - float(client.last_request) > 600:
+                    elif now - float(client.last_request) > 600: # Cette partie du code peut rendre une erreur en essayant de détruire une entité déjà inexistante
                         del_client = db.session.query(Data).filter_by(id=client.id).first()
-                        db.session.delete(del_client)
+                        if del_client is not None: # Vérification du statut de l'entité avant .delete 
+                            db.session.delete(del_client)
             except Exception as e: 
                 print(f"[NESTER][STATUT_MANAGER][ERROR] {e}")
             db.session.commit()
@@ -153,7 +155,7 @@ def connexion():
         else:
             flash('Connexion refusée. Merci de réessayer votre mot de passe !')
             print("Connexion échouée")
-    return render_template('login.html') # Si la requète est "GET" (récupération de donnée )
+    return render_template('login.html', hostname=socket.gethostname()) # Si la requète est "GET" (récupération de donnée )
 
 
 #Voir les infos des clients
@@ -167,7 +169,7 @@ def view_client_info():
 def details(hostname):
     data_details = HarvesterDetails.query.filter_by(hostname=hostname).first()
     if data_details: 
-        return render_template('detail.html', data_details=data_details)
+        return render_template('clientdb.html', data_details=data_details)
     else: 
         return "Aucune information trouvée pour le hostname spécifié.", 404
